+++ classes/class.ilHelpMePlugin.php
@@ -1,15 +1,12 @@
 <?php
 
 require_once __DIR__ . "/../vendor/autoload.php";
-if (file_exists(__DIR__ . "/../../../../Cron/CronHook/HelpMeCron/vendor/autoload.php")) {
-    require_once __DIR__ . "/../../../../Cron/CronHook/HelpMeCron/vendor/autoload.php";
-}
 
 use ILIAS\DI\Container;
 use ILIAS\GlobalScreen\Provider\PluginProviderCollection;
 use srag\CustomInputGUIs\HelpMe\Loader\CustomInputGUIsLoaderDetector;
-use srag\DIC\HelpMe\DevTools\DevToolsCtrl;
-use srag\DIC\HelpMe\Util\LibraryLanguageInstaller;
+use srag\DevTools\HelpMe\DevToolsCtrl;
+use srag\LibraryLanguageInstaller\HelpMe\LibraryLanguageInstaller;
 use srag\Plugins\HelpMe\Utils\HelpMeTrait;
 use srag\RemovePluginDataConfirm\HelpMe\PluginUninstallTrait;